# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RBAC 权限管理系统 API",
    "version": "1.0.0",
    "description": "RBAC权限管理系统后端接口文档",
    "contact": {
      "name": "技术支持",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "获取用户列表",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number",
              "default": 1
            },
            "description": "页码"
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "number",
              "default": 10
            },
            "description": "每页条数"
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            },
            "description": "用户名（模糊搜索）"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "number"
            },
            "description": "状态（0禁用 1启用）"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取列表"
          }
        }
      },
      "post": {
        "summary": "新增用户",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "获取用户详情",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户详情"
          }
        }
      },
      "put": {
        "summary": "编辑用户",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "编辑成功"
          }
        }
      },
      "delete": {
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/users/{id}/roles": {
      "put": {
        "summary": "分配用户角色",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "角色ID列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分配成功"
          }
        }
      }
    },
    "/users/{id}/password": {
      "put": {
        "summary": "重置用户密码",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "新密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置成功"
          }
        }
      }
    },
...
    "/menus/routes": {
      "get": {
        "summary": "获取用户路由菜单",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "获取当前登录用户可访问的路由菜单（用于前端动态路由）",
        "responses": {
          "200": {
            "description": "成功获取用户路由"
          }
        }
      }
    },
...
  },
...
}